HYC-1198 - Update DownloadsController for hyrax 3 (#880)
* Switch downloads controller over to being an override

* Mock GA requests for all tests to catch cases that did not previously make them

* Add routing for Hyrax::ObjectNotFoundError which is now thrown instead of ActiveFedora::ObjectNotFoundError in the hyrax code base (but not active-fedora codebase)

* Removing upstream render_404 method from DownloadBehavior which was causing requests to download files that don't exist to blow up. We had overridden this method previously, but it seems better to fall back to the version from ApplicationController

* Change behavior when access is denied to a file to send users to a 401 rather than 404. Remove handling of RecordNotFound since it is handled by ApplicationControl. Add test coverage for error cases
